\name{TestRcppEigen-package}
\alias{TestRcppEigen-package}
\alias{TestRcppEigen}
\docType{package}
\title{
\packageTitle{TestRcppEigen}
}
\description{
\packageDescription{TestRcppEigen}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TestRcppEigen}
\packageIndices{TestRcppEigen}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TestRcppEigen}

Maintainer: \packageMaintainer{TestRcppEigen}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
